Add `tracing-subscriber` to testing
This should allow us to examine the logs from the roles that are using
`tracing` crate for logging. As a default only ERROR logs will be shown,
in order to examine a different level of logging, `RUST_LOG` env
variable should be modified.

For example `RUST_LOG=info cargo test --test '*'`